requirements.txt
torch
openai-whisper
langchain-community
transformers
huggingface_hub
accelerate
bitsandbytes